+++ src/main/java/imagingbook/pub/edgepreservingfilters/PeronaMalikFilterVector.java
@@ -44,8 +44,8 @@ public class PeronaMalikFilterVector extends GenericFilterVector implements Pero
 	// constructor - use this version to set all parameters
 	public PeronaMalikFilterVector (Parameters params) {
 		this.T = params.iterations;
-		this.alpha = params.alpha;
-		this.g = ConductanceFunction.get(params.conductanceFunType, params.kappa);
+		this.alpha = (float) params.alpha;
+		this.g = ConductanceFunction.get(params.conductanceFunType, (float)params.kappa);
 		this.colorMode = params.colorMode;
 	}
